Initial import:  * ShellCommandExecutorTest
Changeset:
 * Overload the execute method from ShellCommandExecutor in addition to pass command execution timeout different from default one.

[http://jenkins.musala.com/r/8422/] $8422 see # 3133

git-svn-id: https://trac.musala.com/ARDC/repo@10020 34708651-a5da-c34d-b20c-10ddab38317d

+++ src/com/musala/atmosphere/agent/devicewrapper/AbstractWrapDevice.java
@@ -202,7 +202,7 @@ public abstract class AbstractWrapDevice extends UnicastRemoteObject implements 
             // Misc functionalities
             case WAIT_FOR_EXISTS:
                 returnValue = uiAutomatorBridgeCommunicator.waitForExists((UiElementDescriptor) args[0],
-                                                                          (Long) args[1],
+                                                                          (Integer) args[1],
                                                                           wrappedDevice.getSerialNumber());
                 break;
             case PLAY_GESTURE:
